% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dist.deldir.R
\name{as.dist.deldir}
\alias{as.dist.deldir}
\title{Cast deldir Object to Distances Between Neighbours}
\usage{
\method{as.dist}{deldir}(m, diag = FALSE, upper = FALSE)
}
\arguments{
\item{m}{A \code{\link[deldir]{deldir}} result.}

\item{diag}{logical value indicating whether the diagonal of the distance
matrix should be printed by \code{\link{print.dist}}.}

\item{upper}{logical value indicating whether the upper triangle of the
distance matrix should be printed by \code{\link{print.dist}}.}
}
\description{
Function casts a Delaunay triangulation from
\code{\link[deldir]{deldir}} (\pkg{deldir} package) to a
\code{\link{dist}} object, with numeric values for distances
between neighbours and \code{NA} for other distances.
}
