% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.prcomp.rda.R
\name{as.prcomp}
\alias{as.prcomp}
\alias{as.prcomp.rda}
\title{Cast vegan::rda Result to stats::prcomp}
\usage{
as.prcomp(x)

\method{as.prcomp}{rda}(x)
}
\arguments{
\item{x}{An unconstrained \code{\link[vegan]{rda}} result object.}
}
\value{
A \code{\link{prcomp}} result object.
}
\description{
Function casts a result object of unconstrained
\code{\link[vegan]{rda}} or \code{\link[vegan]{pca}} to a
\code{\link{prcomp}} result object, but without the last zero
eigenvalue and rubbish eigenvectors of \code{prcomp}. The
translation is best evaluated by looking at the code of the
function.
}
\author{
Jari Oksanen
}
