% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{bjdbrda}
\alias{bjdbrda}
\alias{scores.bjdbrda}
\alias{plot.bjdbrda}
\alias{boxplot.bjdbrda}
\title{Multiple dbRDA from Beta Distributed Jaccard Dissimilarity}
\usage{
bjdbrda(formula, data, n = 100, ...)

\method{scores}{bjdbrda}(
  x,
  choices = 1:2,
  display = c("wa", "lc", "bp", "cn"),
  scaling = "species",
  const,
  expected = TRUE,
  ...
)

\method{plot}{bjdbrda}(
  x,
  choices = 1:2,
  wa = "p",
  lc = "n",
  cn = "hull",
  bp = "wedge",
  wa.par = list(),
  lc.par = list(),
  cn.par = list(),
  bp.par = list(),
  scaling = "species",
  type = "t",
  ...
)

\method{boxplot}{bjdbrda}(
  x,
  kind = c("eigen", "correlation"),
  points = "red",
  pch = 16,
  xlab = "Axis",
  ylab,
  ...
)
}
\arguments{
\item{formula, data}{Model definition of type \code{Y ~ Var1 + Var2,
data = X}, where \code{Y} is dependent community data (handled
as binary data), \code{Var1} and \code{Var2} are independent
(explanatory) variables found in data frame \code{X}. See
\code{\link[vegan]{dbrda}} for further information.}

\item{n}{Number of Jaccard dissimilarity matrices sampled from Beta
distribution.}

\item{...}{Other parameters passed to functions; passed
\code{\link[vegan]{dbrda}} in \code{bjdbrda}, ignored in
\code{scores}.}

\item{x}{\code{bjdbrda} result object.}

\item{choices}{Selected ordination axes.}

\item{display, scaling, const}{Kind of scores, the scaling of scores
(axes), and scaling constant with similar definitions as in
\code{\link[vegan]{scores.rda}}.}

\item{expected}{Return scores of the expected ordination instead of
ordination based on random samples from Jaccard dissimilarity.}

\item{wa, lc, cn, bp}{Display of corresponding scores. \code{"n"}
skips the score, \code{"p"} and \code{"t"} use points or text
for the expected score, and other shapes define
\code{\link{bjpolygon}} or \code{\link{bjstars}} shape used for
sampled random scores.}

\item{wa.par, lc.par, cn.par, bp.par}{List of arguments to modify the
plotting parameters of the corresponding shape.}

\item{type}{Add \code{"t"}ext or \code{"p"}oint for the expected
score to a shape of scatter of random shapes.}

\item{kind}{Draw boxplots of eigenvalues or pairwise axis correlations
for each random sample.}

\item{points, pch}{Add points of given color and shape to the
eigenvalue boxplot.}

\item{xlab, ylab}{Change labelling of axes in boxplot.}
}
\value{
Function returns \code{\link{dbrda}} result object amended
    with item \code{BayesJaccard} that is a list of randomized
    results with following items for each random sample:
\itemize{
  \item \code{tot.chi} total inertia
  \item \code{eig} eigenvalues of constrained axes
  \item \code{r} absolute value of correlation coefficient of
    randomized axis and the reference axis
  \item \code{u, wa, biplot, centroids} ordination scores for linear
    combination scores, weighted averages scores, biplot scores and centroid
    of factor constraints; these are unscaled and are usually accessed with
    \code{\link{scores.bjdbrda}} for appropriate scaling
}
}
\description{
Function performs distance-based RDA on expected Beta Jaccard
dissimilarities, and then reruns the analysis on given number of
random Beta Jaccard dissimilarities.
}
\details{
Function works only with constrained and partial constrained
ordination, and only saves the randomized results of constrained
ordination. To maintain the eigenvalues, function does not rotate
the random results to the reference ordination, but it will reflect
axes with reversed signs. The eigenvalues and magnitudes of
axiswise correlations to the reference ordination can be inspected
with \code{boxplot}.

The display type in \code{plot} can be specified independently to
each type of score (or the score can be skipped). The default
display can be modified using a similarly named list of graphical
argument values that will replace the defaults. For instance, to
display linear combination scores as convex hull, use \code{lc =
"hull"}, and modify its parameters with list \code{lc.par}.
}
\examples{
if (require(vegan)) {
data(dune, dune.env)
m <- bjdbrda(dune ~ A1 + Management + Moisture, dune.env)
## eigenvalues and correlations showing axis stability
boxplot(m)
boxplot(m, kind = "correlation")
## Default plot
plot(m)
## modify plot
plot(m, wa = "ellipse",
   wa.par=list(col = dune.env$Management, keep = 0.607))
}

}
