% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{bjpolygon}
\alias{bjpolygon}
\alias{bjstars}
\title{Shapes to Display Scatter of Points in Multiple Ordinations}
\usage{
bjpolygon(
  xarr,
  x0,
  keep = 0.9,
  kind = c("hull", "ellipse"),
  criterion = "area",
  linetopoint = TRUE,
  col = "gray",
  alpha = 0.3,
  observed = TRUE,
  type = c("t", "p", "n"),
  ...
)

bjstars(xarr, x0, keep = 0.9, col = "gray", type = c("t", "p", "n"), ...)
}
\arguments{
\item{xarr}{3-D array of coordinates of sampling units times two
axes by random samples.}

\item{x0}{2-D array of sampling units times two axes treated as
constant for all samples in \code{xarr}.}

\item{keep}{Proportion of points enclosed in the shape; passed to
\code{\link{peelhull}} or \code{\link{peelellipse}}.}

\item{kind}{Shape is either a convex hull (\code{\link{peelhull}})
or ellipse \code{\link{peelellipse}} enclosing \code{keep}
proportion of \code{xarr} points.}

\item{criterion}{Criterion to remove extreme points from the hull
in \code{\link{peelhull}} and \code{\link{peelellipse}}.}

\item{linetopoint}{Draw line from the centre of the shape to the
coordinates in \code{x0}.}

\item{col}{Colour of the shape; can be a vector of colours.}

\item{alpha}{Transparency of shapes; 0 is completely transparent,
and 1 is non-transparent.}

\item{observed}{After finding the convex hull, extend hull to
enclose fixed point \code{x0}.}

\item{type}{Mark coordinate of \code{x0} using \code{"t"}ext,
\code{"p"}oint or \code{"n"}one.}

\item{...}{Other parameters passed to to marker of \code{type}.}
}
\description{
Shapes to Display Scatter of Points in Multiple Ordinations
}
