% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMDS.R
\name{cNMDS}
\alias{cNMDS}
\title{Constrained Nonmetric Multidimensional Scaling}
\usage{
cNMDS(formula, data, k = 2)
}
\arguments{
\item{formula}{Model formula where the left-hand side must be
dissimilarities, and right-hand side the constraints.}

\item{data}{Data frame to find the terms on the right-hand side of
the formula.}

\item{k}{Number of dimensions in NMDS.}
}
\description{
Adds points based on community dissimilarities on the NMDS
based on constrained dissimilarities.
}
\details{
The steps of algorithm are:
\enumerate{
  \item Find constrained dissimilarities for given constraints
    with \code{\link{distconstrain}}.
  \item Find NMDS of constrained dissimilarities with
    \code{\link[vegan]{monoMDS}} using metric scaling as initial
    configuration.
  \item Add points to constrained ordination using unconstrained
    community dissimilarities with \code{\link[vegan]{MDSaddpoints}}.
} % end enumerate
}
