% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distconstrain.R
\name{distconstrain}
\alias{distconstrain}
\title{Constrained and Residual Dissimilarities}
\usage{
distconstrain(formula, data, add = FALSE, residuals = FALSE,
  squared = FALSE)
}
\arguments{
\item{formula}{The left-hand-side must be dissimilarities and the
right-hand-side should list the constraining variables.}

\item{data}{Data frame containing the constrainging variables in
the \code{formula}.}

\item{add}{an additive constant is added to the non-diagonal
dissimilarities such that all \eqn{n-1} eigenvalues are
non-negative. Alternatives are \code{"lingoes"} (default, also
used with \code{TRUE}) and \code{"cailliez"} (which is the only
alternative in \code{\link{cmdscale}}).}

\item{residuals}{Return residuals after constraints.}

\item{squared}{Return squared dissimilarities instead of
dissimilarities. This allows handling negative squared distances by
the user instead of setting them \code{NaN}.}
}
\description{
Function constrains dissimilarities by external variables, or
alternatively removes effects of constraining variables and returns
residual dissimilarities. The analysis is based on McArdle &
Anderson (2001), and the analysis of constrained dissimilarities is
equal to distance-based Redundancy Analysis
(\code{\link[vegan]{dbrda}}).
}
\details{
Function uses the method of McArdle & Anderson (2001) to
constrain dissimilarities by external variables, or alternatively,
to find residual dissimilarities after constraints. With Euclidean
distances, the method is equal to performing linear regressions on
each column in the raw data and then calculationg the distances,
but works directly on distances. With other methods, there is no
similar direct connection to the raw data, but it is possible to
work with non-Euclidean metrics. The same basic method is used
within db-RDA (\code{\link[vegan]{dbrda}} in \CRANpkg{vegan}), but
this function exposes the internal calculations to users.

Non-Euclidean indices can produce negative eigenvalues in
db-RDA. Would negative eigenvalues be produced, this function can
return negative squared distances resulting in \code{NaN} when
taking the square root. Db-RDA works with the internal presentation
of the dissimilarities, and its analysis does not suffer from the
imaginary distances, but these can ruin the analysis of
dissimilarities returned from this function.
}
\references{
McArdle, B.H. & Anderson, M.J. (2001). Fitting
  multivariate models to community data: a comment on distance-based
  redundancy analysis. \emph{Ecology} 82, 290--297.
}

