% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMeans.R
\name{distMeans}
\alias{distMeans}
\title{Mean of Distances}
\usage{
distMeans(d, addcentre = FALSE, label = "centroid")
}
\arguments{
\item{d}{Distances as a \code{dist} object}

\item{addcentre}{Add distances to the centroid as the first item in
the distance matrix. If \code{FALSE} only return mean
distances.}

\item{label}{Label for the centroid when \code{addcentre = TRUE}.}
}
\value{
Either istances to all other points from a point that is in
    the centroid of the coordinates generating the distances, or
    the input dissimilarity matrix where the mean distances are
    added as the first observation.
}
\description{
Mean of distances is defined as the distance of each point to the
mean of coordinates generating the distance.
}
