% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diverclust.R, R/diverdist.R
\name{diverclust}
\alias{diverclust}
\alias{diverdist}
\title{beta and gamma Diversity Clustering}
\usage{
diverclust(
  x,
  renyi = 1,
  equalize = c("renyi", "total", "no"),
  beta = TRUE,
  hill = FALSE,
  trace = FALSE
)

diverdist(
  x,
  renyi = 1,
  equalize = c("renyi", "total", "no"),
  beta = TRUE,
  hill = FALSE
)
}
\arguments{
\item{x}{Community data.}

\item{renyi}{The scale of Renyi diversity index as defined in
\code{\link[vegan]{renyi}}. Value \code{0} gives diversity based on
species richness, \code{1} diversity based on Shannon entropy,
\code{2} diversity based on Simpson index, and \code{Inf} diversity
based on Berger-Parker index. Other non-negative values are also
allowed, but they may not define a well known standard diversity
index.}

\item{equalize}{Equalize data rows so that they can be meaningfully
pooled and averaged. If this is \code{FALSE}, raw data will be
used, and beta diversities may be negative. The equalization
depends on the value of \code{renyi}, and each row of \code{x} is
divided by \code{(rowSums(x^renyi))^(1/renyi)}. In borderline cases
\code{renyi=0} the data are presence-absence tranformed (but can be
left untransformed), and with \code{renyi=Inf} the rows are divided
by row maxima (see \code{\link[vegan]{decostand}}).}

\item{beta}{Use beta diversities: the average alpha diversity of
cluster members is subtracted from the pooled diversity. If this is
\code{FALSE},the clustering is based on pooled diversity, also
known as gamma diversity.}

\item{hill}{Use Hill numbers instead of Renyi diversity (see
\code{\link[vegan]{renyi}}). For \code{renyi = 0} these are species
richess values instead of their logarithms, and for \code{renyi =
1} they are exponents of Shannon diversity instead of Shannon
diversities. In general, a Hill number is an exponent of Renyi
diversity. The Hill numbers may not be strictly additive, and beta
diversities may be negative (except with \code{renyi = 0}).}

\item{trace}{Trace calculations. Either logical or an integer:
\code{trace=2} also traces merges.}
}
\value{
\code{diverclust} returns an \code{\link{hclust}} object.

\code{diverdist} returns a dissimilarity object inheriting
    from \code{\link{dist}}.
}
\description{
\code{diverclust} forms hierarchic clustering so that each fusion
minimizes beta diversity or gamma diversity.

The distance between two sampling units is the increase in
deversity when these are merged together, or the beta diversity
between fused units. \code{diverdist} finds such distances for all
pairs of sampling units using any Rényi diversity.
}
\details{
\code{diverclust} forms clusters so that pooled diversity or
its change to the baseline alpha diversity is minimized at each
level of clustering. The change of pooled diversity with respect to
the baseline alpha diversity is called beta diversity, and the
overall diversity is called gamma diversity. For beta diversity,
the clustering implies an additive partitioning.

Function \code{diverdist} finds beta diversities (or optionally
gamma diversities) for pairs of sampling units. These are used as
starting values in \code{diverclust}, but after merging sampling
units into clusters, beta diversities are calculated for clusters
and cannot be found directly from these dissimilarities.

Both \code{diverdist} and \code{diverclust} work by pooling
together sampling units and averaging species abundances. The units
may be equalized before pooling together to remove the effects of
variable totals in sampling units. Rényi diversities are based on
species proportions and most popular equalization is to use species
proportions by sampling units instead of raw data (argument
\code{equalize="total"}). With alternative \code{"renyi"}
equalization depends on the\code{renyi} scale. With \code{renyi=2}
(Simpson type) sums of squared proportions or norms are made equal,
and with \code{renyi=Inf}, species maxima are made equal in
compared units, and with \code{renyi=1} (Shannon type),
\code{"total"}. Equalization has no effect with \code{scale=0}
(Species richness type).

For more detailed description, see vignette \code{"diverclust"}.
}
\examples{
data(spurn)
(cl <- diverclust(spurn))
plot(cl) # beta diversity based on Shannon diversity
(cl <- diverclust(spurn, renyi=0, hill=TRUE))
if (require(vegan, quietly=TRUE))
   tabasco(spurn, cl)
plot(cl) # increase in species richness when pooling sites

## increase in Shannon diversity when pooling two sites
d <- diverdist(spurn, renyi=1, equalize="no")  # raw data
de <- diverdist(spurn, renyi=1, equalize="total")
plot(d, de, xlab = "Shannon distance, no equalization",
            ylab = "Shannon distance, equalized")

}
\seealso{
\code{\link{hclust}} for cluster analysis and its support
    methods, \code{\link[vegan]{renyi}} for estimating Renyi
    diversities and Hill numbers, and \code{\link[vegan]{adipart}}
    for related additive partitioning of beta diversity.
}
\author{
Jari Oksanen.
}
