% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gausscircle.R
\name{gausscircle}
\alias{gausscircle}
\alias{gaussellipse}
\title{Circular Gaussian Response on 2D Ordination}
\usage{
gausscircle(
  ord,
  comm,
  freqlim = 10,
  family = quasipoisson(),
  unit = FALSE,
  choices = 1:2,
  display = "sites",
  ...
)

gaussellipse(
  ord,
  comm,
  freqlim = 10,
  family = quasipoisson(),
  choices = 1:2,
  display = "sites",
  ...
)
}
\arguments{
\item{ord}{Ordination result.}

\item{comm}{Community data.}

\item{freqlim}{Frequency limit.}

\item{family}{Error family.}

\item{unit}{Force Gaussian responses to unit tolerances.}

\item{choices}{Ordination axes.}

\item{display}{Ordination scores.}

\item{\dots}{Other arguments passed to \code{\link{scores}}.}
}
\description{
Circular Gaussian Response on 2D Ordination
}
\author{
Jari Oksanen
}
