% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginkgogram.R
\name{ginkgogram}
\alias{ginkgogram}
\title{Plot Dendrogram with Leaves as Fans}
\usage{
ginkgogram(x, labels = NULL, check = TRUE, axes = TRUE,
  frame.plot = FALSE, ann = TRUE, main = "Cluster Ginkgogram",
  sub = NULL, xlab = NULL, ylab = "Height", w, col, ...)
}
\arguments{
\item{x}{An object of the type produced by \code{\link{hclust}}.}

\item{labels}{A character vector of labels for te leaves of the tree.}

\item{check}{Logical indicating if the \code{x} should be checked
for validity.}

\item{axes, frame.plot, ann}{Logical flags as in \code{\link{plot.default}}.}

\item{main, sub, xlab, ylab}{Character strings to replace default annotation.}

\item{w}{Weights of leaves.}

\item{col}{Colour of the fill of leaves.}

\item{\dots}{Further graphical arguments}
}
\description{
Function is similar to standard \code{\link{hclust}}, but the
leaves are drawn as fans with base proportional to weights (or
sizes) of leaves.
}
