% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradrescale.R
\name{gradrescale}
\alias{gradrescale}
\title{Rescale Ecological Gradient to Constant Community Heterogeneity}
\usage{
gradrescale(grad, comm)
}
\arguments{
\item{grad}{Observed numeric gradient.}

\item{comm}{Community data matrix.}
}
\value{
Rescaled gradient scaled in units of heterogeneity.
}
\description{
Rescale Ecological Gradient to Constant Community Heterogeneity
}
\details{
Observed environmental gradient is rescaled similarly as axis in
\code{\link{rdecorana}} ahd \code{\link[vegan]{decorana}}. The goal
is that weighted averages of species have equal dispersion along
the rescaled gradient.
}
\examples{
if(require(vegan, quietly=TRUE)) {
data(mite, mite.env)
WatrCont.rescaled <- with(mite.env, gradrescale(WatrCont, mite))
plot(WatrCont.rescaled ~ WatrCont, data=mite.env)
}

}
\author{
Jari Oksanen
}
