% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iap.R
\name{iapq}
\alias{iapq}
\alias{iap}
\alias{plot.iapq}
\alias{summary.iapq}
\title{Number of Companion Species (IAP)}
\usage{
iapq(comm, freq.min = 5, permutations = 999)

iap(comm, iapq)

\method{plot}{iapq}(x, type = c("p", "t"), optimize = FALSE, bg, labels, cex = 0.7, ...)

\method{summary}{iapq}(object, ...)
}
\arguments{
\item{comm}{The community data frame.}

\item{freq.min}{Minimum number of occurrences for analysed species.}

\item{permutations}{Number of permutations to assess the randomized
number of companion species.}

\item{iapq}{Result of \code{iapq}.}

\item{x}{\code{iapq} result object.}

\item{type}{Plot \code{"p"}oints or \code{"t"}ext.}

\item{optimize}{Optimize location of points to minimize overplotting.}

\item{bg}{Background colour of labels for text.}

\item{labels}{Text labels to replace original species names.}

\item{cex}{Character expansion for text and points.}

\item{\dots}{Other arguments to the function.}

\item{object}{\code{iapq} result object.}
}
\description{
Function finds the number of companion species or the average
species richness of other species for each species in a
community. This is the quality index \eqn{Q} of Index of
Atmospheric Purity (IAP) in lichen bioindication (LeBlanc & De
Sloover 1970). The non-randomness of low or high \eqn{Q} values is
found by randomization.
}
\details{
Index of Atmospheric Purity (IAP) is used in bioindication
with epiphytic lichens and bryophytes (LeBlanc & De Sloover
1970). It derives species indicator scores (\eqn{Q}) as the number
of other species in sampling units where each focal species is
present, and then finds the IAP values for each sampling unit as
scaled weighted sum of species indicator values.

Function \code{iapq} finds the \eqn{Q} values for all species in a
community data set. Function \code{iap} applies these values for a
community data set to evalute the IAP values for each site. The
species are matched by names. LeBlanc & De Sloover (1970) used
scaled abundance values and divided the weighted sum by 10, but
this is not done in the current function, but this is left to the
user.

Function \code{iapq} is a general measure of indicator value for
species richness and can well be used outside lichen
bioindication. The \eqn{Q} value is the average species richness in
sampling units where the species is present, excluding the species
itself from the richness. For rare species, \eqn{Q} is based on
small sample size, and is therefore more variable than for common
species. The \code{iapq} function assesses the non-randomness
(\sQuote{significance}) of \eqn{Q} by taking random samples of the
same size as the frequency (number of occurrence) of the focal
species and finding the average richness (without the focal
species) in these samples. Because species are more likely to be
present in species-rich sampling units than in species-poor, the
random sampling uses the observed species richness (with the
focal species) as weights in random sampling. Testing is two-sided
and the number of greater or less random values is multiplied with
two. The observed value of \eqn{Q} is included in the random sample
of species richness values both in assessing the \eqn{p}-value and
in estimating the quantiles.
}
\examples{
data(spurn)
iq <- iapq(spurn)
plot(iq, type="t", optimize = TRUE, bg = "white")
iap(spurn, iq)

}
\references{
LeBlanc, S.C. & De Sloover, J. (1970) Relation between
industrialization and the distribution and growth of epiphytic
lichens and mosses in Montreal. \emph{Can. J. Bot.} 48, 1485--1496.
}
