% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoclust.R
\name{infoclust}
\alias{infoclust}
\title{Information Clusterings of Ecological Communities}
\usage{
infoclust(x, delta = TRUE)
}
\arguments{
\item{x}{Community data}

\item{delta}{Use increase in information (\eqn{\Delta I}) instead of the
value of information (\eqn{I}) when merging clusters as
recommended by Williams et al. (1966) and Lance & Williams (1966).}
}
\value{
Function returns an object of class \code{"infoclust"} that
inherits from \code{\link{hclust}}. It uses all \code{"hclust"}
methods, but some may fail or work in unexpected ways because the
analysis is not based on dissimilarities but on binary data matrix.
}
\description{
Function performs hierarchical clustering of binary ecological
communities based on information analysis as defined by Williams et
al. (1966) and Lance & Williams (1966).
}
\details{
Function performs information analysis of binary
ecological communities (Williams et al. 1966, Lance & Williams
1966). The current implementation is based on Legendre & Legendre
(2012).

The information \eqn{I} of a collection of \eqn{N} sampling units
with \eqn{S} species is defined as
  \deqn{I = S N \log N - \sum_i^S a_i \log a_i + (N-a_i) \log (N-a_i)}{I = S*N*log(N) - sum(a*log(a) + (N-a)*log(N-a))}
where \eqn{a} is the frequency count of each species in the
collection. The method works by merging either the units that give
the lowest increase (\eqn{\Delta I} when \code{delta = TRUE}), or
the units that are most homogeneous (lowest \eqn{I} when
\code{delta = FALSE}). After merging sampling units or clusters,
the community data matrix is updated by actually merging the data
units and re-evaluating their information distance to all other
units. The information content of all non-merged clusters is \eqn{I
= 0}, and for clusters of several sampling units the constant
species (completely absent or always present) do not contribute to
the information. The largest increase in information is made by
species with 0.5 relative frequency, so that the analysis tries to
build clusters where species is either always present or always
absent. This often gives easily interpretable clusters.
}
\examples{
## example used to demonstrate the calculation of
## information analysis by Legendre & Legendre (2012, p. 372).
data(pond)
cl <- infoclust(pond)
plot(cl, hang = -1)
## Lance & Williams suggest a limit below which clustering is
## insignificant and should not be interpreted
abline(h=qchisq(0.95, ncol(pond)), col=2)

}
\references{
Williams, W.T., Lambert, J.M. & Lance,
G.N. (1966). Multivariate methods in plant ecology. V. Similarity
analyses and information-analysis. \emph{J. Ecol.} 54, 427--445.

Lance, G.N. & Williams, W.T. (1966). Computer programs for
hierarchical polythetic classification (\dQuote{similarity
analyses}). \emph{Comp. J.} 9, 60--64.

Legendre, P. & Legendre, L. (2012). \emph{Numerical Ecology.} 3rd
English Ed., Elsevier.
}
