% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mrankdist.R
\name{mrankdist}
\alias{mrankdist}
\title{Mean Rank Shift Between Pairs of Sites}
\usage{
mrankdist(x, missing = TRUE)
}
\arguments{
\item{x}{The input data.}

\item{missing}{If \code{TRUE}, missing species (zero abundance)
will not be used in comparing rank shifts, but rank shifts are only
compared for species that occur in both compared SUs. If
\code{FALSE}, the missing species are given tied rank as last
species.}
}
\description{
Function calculates the mean rank shift between decreasing rank
orders of species (columns). The averaging is done by the number of
species occurring in both compared sites, and the results are
returned as dissimilarities.
}
\references{
Collins, S.L, Suding, K.N., Cleland, E.E., Batty, M.,
Pennings, S.C., Gross, K.L., Grace, J.B., Gough, L., Fargione,
J.E. & Clark, C.M. (2008). Rank clocks and plant community
dynamics. \emph{Ecology} 89, 3534--3541.
}

