\name{mtfield5}
\alias{mtfield5}
\docType{data}

\title{
Occurrences of Five Plant Species and Altitude on Mt Field, Tasmania
}

\description{

  Occurrences of five plant species on 167 sample plots on Mt Field,
  Tasmania, and the Altitude above sea level (in meters) of the sample
  plots. The data are the well drained subset of the complete of
  Minchin (1989). These data were also analysed by Oksanen & Minchin
  (2002). The species are \emph{Epacris serpyllifolia} (abbreviated as
  \code{Epacserp}), \emph{Cyathodes petiolaris} (\code{Cyatpeti}),
  \emph{Nothofagus cunninghamiana} (\code{Nothcunn}), \emph{Poa
  gunnii} (\code{Poagunn}), and \emph{Bauera rubioides}
  (\code{Bauerubi}).

}

\usage{data("mtfield5")}

\references{
  Minchin, P.R. (1989). Montane vegetation of the Mt. Field 
  massif, Tasmania: a test of some hypotheses about properties of
  community patterns. \emph{Vegetatio} 83, 97--110.

  Oksanen, J. & Minchin, P.R. (2002). Continuum theory revisited:
  what shape are species responses along ecological gradients?
  \emph{Ecological Modelling} 157, 119--129.
}
