% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natto-package.r
\docType{package}
\name{natto}
\alias{natto-package}
\alias{natto}
\title{natto.}
\description{
Random code that is too experimental or too weird to be included in the vegan package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jarioksa/natto}
}

}
\author{
\strong{Maintainer}: Jari Oksanen \email{jhoksane@gmail.com}

}
