% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho.R
\name{ortho}
\alias{ortho}
\title{Gram-Schmidt Orthogonalization}
\usage{
ortho(x0, x)
}
\arguments{
\item{x0}{Matrix against which orthogonalization is calculated.}

\item{x}{vector to be orthogonalized.}
}
\description{
Function orthogonalizes vector \code{x} against all columns of
matrix \code{x0}.
}
