% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarord.R
\name{polarord}
\alias{polarord}
\alias{plot.polarord}
\alias{sppscores<-.polarord}
\alias{sppscores<-}
\alias{eigenvals.polarord}
\alias{eigenvals}
\title{Polar (Bray-Curtis) Ordination}
\usage{
polarord(d, k = 2, endpoints)

\method{plot}{polarord}(x, choices = c(1, 2), type = "t", display, ...)

\method{sppscores}{polarord}(object) <- value

\method{eigenvals}{polarord}(x, ...)
}
\arguments{
\item{d}{Dissimilarities or distances: a \code{\link{dist}} object.}

\item{k}{Number of dimensions.}

\item{endpoints}{Indices (not names) of endpoints for axis
one. This can be either a single integer for the first
endpoint, and the second endpoint is found automatically, or a
vector of two endpoints.}

\item{x}{\code{polarord} result.}

\item{choices}{Axes shown.}

\item{type}{Type of graph which may be \code{"t"} for text, \code{"p"}
for points or \code{"n"} for none (an empty plot).}

\item{display}{Items displayed: \code{"sites"} are always available,
but \code{"species"} only if they were added with sppscores.}

\item{\dots}{Other arguments to the function (passed to
\code{\link[vegan]{ordiplot}}).}

\item{object}{\code{polarord} result.}

\item{value}{Community data to find the species scores.}
}
\value{
The function returns an object of class \code{"polarord"} with the
following elements:
\itemize{
  \item \code{points}: The ordination scores.
  \item \code{inertia}: Total inertia of the input dissimilarities.
  \item \code{eig}: Eigenvalues of axes. These do not usually add up to
     total inertia and may not be in strictly descending order.
  \item \code{endpoints}: The indices (not the names) of the endpoints for
     each axis.
  \item \code{call}: Function call.
}
}
\description{
Polar or Bray-Curtis ordination is a historic ordination method
that could be performed without computers with simple hand
calculations (Bray & Curtis 1957). Ordination axis is found by
selecting two extreme points and projecting all points between
these end points. The current function follows Beals (1984) in
selecting the endpoints, projection of points on axis, and defining
the residual distances for later axes.
}
\details{
The implementation follows McCune & Grace (2002, Chapter 17). The
endpoints are found using the variance-regression method of Beals
(1984). The first endpoint has the highest variance of distances to
other points. This guarantees that the point is at the margin of
the multivariate cloud but is not an outlier, since outliers have
long distances to all points and hence low variance of
distances. The second endpoint has the lowest (most negative)
regression coefficient between distances from the first and second
point to all other points. This selects a point at the margin of
the main cloud of points, opposite to first endpoint. All points
are projected on the axis between the endpoints using Euclidean
geometry, and this gives the scores on a polar ordination
axis. Then the effect of the axis are removed by calculating
residual distances using Euclidean geometry. Ecological indices are
usually semimetric, and negative residual distances can emerge, but
these are taken as zero in the current function.

The eigenvalues are estimated from the dispersion of points, and
they are not necessarily in descending order. Usually only some
first axes are reliable, and too high numbers of dimensions should
be avoided. The inertia and eigenvalues give the average
dispersion, and they are consistent with \code{\link[vegan]{dbrda}}
and \code{\link[vegan]{capscale}}.

Polar ordination is a historical method that is little used today,
but several authors claim that it is a powerful method (see McCune
& Grace 2002). Although the basic operations can be easily
performed by hand or graphically, the later developments of
endpoint selection require more extensive calculations. With modern
numerical utilities, the polar ordination is not faster than metric
multidimensional scaling (\code{\link{cmdscale}},
\code{\link[vegan]{wcmdscale}}).

It is possible to use predefined endpoints for axis 1 instead of
automatic selection. This can be useful for confirmatory analysis
(McCune & Grace 2002). If only one endpoint is given, the second
one is chosen automatically.

Function \code{sppscores} can be used to add species scores
  to the ordination result.
}
\examples{

data(spurn)
dis <- dist(spurn, method = "binary") ## Jaccard index
ord <- polarord(dis)
ord
summary(eigenvals(ord))
## add species scores
sppscores(ord) <- spurn
plot(ord)

}
\references{
Beals, E. W. (1984) Bray-Curtis ordination: an effective strategy
for analysis of ecological multivariate data. \emph{Advances in
Ecological Research} 14, 1--55.

Bray, J. R. & Curtis, J. T. (1957) An ordination of the upland
forest communities in southern Wisconsin. \emph{Ecological
Monographs} 27, 325--349.

McCune, B. & Grace, J. B. (2002) \emph{Analysis of Ecological
Communities.} MjM Software Design.
}
