\name{pond}
\alias{pond}
\docType{data}
\title{
Relative Zooplankton Abundances in Ponds
}
\description{
Relative abundances in scale 0 to 5 for un-named zooplankton species in
ponds. 
}
\usage{data("pond")}
\format{
  A data frame with 5 ponds (observations) on 8 species (variables).
}
\details{
  This small data set was used to demonstrate dissimilarity and
  classification methods by Legendre & Legendre (2012).
}
\source{
  Legendre, P. & Legendre, L. (2012). \emph{Numerical Ecology}. 3rd
  English Ed. Elsevier, page 290.
}
\examples{
data(pond)
}
\keyword{datasets}
