% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posvectord.R
\name{posvectord}
\alias{posvectord}
\alias{spvectord}
\alias{plot.posvectord}
\alias{sppscores<-.posvectord}
\alias{scores.posvectord}
\title{Position Vector and Species Vector Ordination}
\usage{
posvectord(x, scale = FALSE)

spvectord(x, scale = FALSE)

\method{plot}{posvectord}(x, choices = c(1, 2), type = "t", display = c("sites", "species"), ...)

\method{sppscores}{posvectord}(object) <- value

\method{scores}{posvectord}(x, display = c("sites", "species"), choices = 1:2, ...)
}
\arguments{
\item{x}{Input data.}

\item{scale}{Scale variables to unit variance.}

\item{choices}{Axes shown.}

\item{type}{Type of plot: \code{"t"} for text, \code{"p"} for
points and \code{"n"} for none.}

\item{display}{Scores shown.}

\item{\dots}{Other arguments (passed to \code{\link[vegan]{ordiplot}}).}

\item{object}{Ordination object to which species scores should be added.}

\item{value}{Community for which species scores should be added.}
}
\value{
\code{posvectord} returns an object of class
    \code{"posvectord"}, and \code{spvectord} returns an object of
    class \code{"spvectord"} that inherits from
    \code{"posvectord"}. Result objects have the following
    elements:

\itemize{
   \item \code{points}: The ordination scores. In SVO, these are
    named by the species (variable) the axes is based on, the
    numerical scores are the centred (residual) values of
    observed data. In PVO, they are called as PVO, and they are
    scaled similarly as in PCA and can reconstitute
    (a low-rank approximation of) covariances.
   \item \code{totvar}: The total variance in the input data.
   \item \code{eig}: Eigenvalues of axes.
   \item \code{posits}: Indices of points used to position axes in
     \code{posvectord}.
}
}
\description{
Position Vector Ordination (PVO) finds ordination axes that go
through sample points in species space and maximize the projections
of other points onto them. These are similar to Principal
Components. Species Vector Ordination (SVO) finds a set of species
that maximize the projections of other species on species
axes. This also can be similar to Principal Components and can be
used to select a subset of species that cover most of the variance
in the data.
}
\details{
Position Vector Ordination (PVO, function \code{posvectord}) is a
simple educational method that resembles Principal Component
Analysis, PCA (Orlóci 1966, 1973a). Function \code{posvectord}
positions vectors from the data centroid (origin) to sample points
in species space and evaluates the projections of other sample
vectors on these positioned axes, and selects the one with highest
total projection as the ordination axis. Then the effects of that
selected axis are removed from the covariance matrix, zeroing the
row and column of selected sampling unit, and the process is
repeated. Principal Components maximize this projection, but PVO
axes can be close to the Principal Component, in particular in the
large data sets with many observations. The method was proposed as
a computationally light approximation to PCA suitable for the
computers of 1960s (Orlóci 1966). Now it can only be regarded as
historically interesting, and also as an educational tool in
introducing PCA.

Species Vector Ordination (SVO; function \code{spvectord}) is
similar, but it picks the species vector that maximizes the species
projected onto that vector (Orlóci 1973b).  The method picks the
species or variable that explains largest proportion of variance
and uses the centred values of this variable as the axis. Then it
orthogonalizes all species or variables to that selected axis and
repeats the selection process. PCA finds a linear combination of
species or variables that minimizes the residual variance, but in
SVO only one species or variable is used. The axes are named by the
species or variables, and the axis scores are the centred
(residual) observed values. The resulting plot has orthogonalized
set of species as axes. SVO can be similar to PCA, in particular
when some few species contribute most of the the total
variance. The method only has educational use in explaining PCA in
species space. Orlóci (1973b) suggesed SVO as a method of selecting
a subset of species or variables that contributes most of the
variance in the data. The current \code{spvectord} function can
also be used for this purpose, although it is mainly geared for
ordination. \CRANpkg{dave} package has function \code{orank}
specifically written for variable or species selection using the
same algorithm.

Function \code{sppscores} can add species scores to the ordination
results. The species scores should be regarded as arrow heads of
vectors. Although the axes of \code{spvectord} are the species, the
corresponding arrow is not usually parallel to the axis after axis
1. the axis shows the residual abundance of the species after
previous axes whereas the species score shows the correlated
response with other species.
}
\examples{
data(spurn)
m <- spvectord(spurn)
sppscores(m) <- spurn
plot(m)
m <- posvectord(spurn)
m
plot(m)
if (require(vegan, quietly = TRUE)) {
## similar to PCA
plot(procrustes(pca(spurn), m, choices=1:2))
}
}
\references{
Orlóci, L. (1966) Geometric models in ecology. I. The theory and
application of some ordination methods. \emph{J. Ecol.} 54: 193--215.

Orlóci, L. (1973a) Ordination by resemblance matrices. In:
R. H. Whittaker (ed.) \emph{Ordination and Classification of
Communities. Handbook of Vegetation Science} 5: 249--286.

Orlóci, L. (1973b) Ranking characters by dispersion
criterion. \emph{Nature} 244: 371--373.
}
\seealso{
\code{\link{polarord}} (Polar Ordination) is a similar
    educational tool to approximate Principal Coordinates Analysis
    (PCoA).
}
\author{
Jari Oksanen
}
