% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projectedmap}
\alias{projectedmap}
\title{Return a projected map with given extent}
\usage{
projectedmap(extent, CRS = "+proj=longlat +datum=WGS84",
  inCRS = "+proj=longlat +datum=WGS84", resolution = "low", pad = 0.04)
}
\arguments{
\item{extent}{Either a four element numeric vector of east min,
east max, north min, north max or a two-column matrix of
coordinates of points that should be shown on the map.}

\item{CRS}{CRS string defining the output projection.}

\item{inCRS}{CRS string describing the input projection of
\code{extent}. This will not be used if CRS is already defined for
the \code{extent}.}

\item{resolution}{Resolution of the map as defined in
\code{\link[rworldmap]{getMap}}.}

\item{pad}{extend climits by this proportion of coordinate range.}
}
\description{
The function takes either limits of extent or a two-column matrix
and returns a projected map containing the original limits or the
points.
}

