% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raostand.R
\name{raostand}
\alias{raostand}
\title{Standardize Data to Yield Rao Satistics}
\usage{
raostand(x, d, propx = TRUE, dmax)
}
\arguments{
\item{x}{Community data.}

\item{d}{Phylogenetic distances or other dissimilarities among
species.}

\item{propx}{Standardize row of \code{x} to unit total. This will
give standardization that is compatible with \code{qrao} and
\code{distrao}.}

\item{dmax}{Scale dissimilarities by \code{dmax} (if \code{dmax >
max(d)}) or truncate dissimilarities at \code{dmax} (if
\code{dmax < max(d)}).}
}
\description{
Function standardizes data so that it gives Rao phylogenetic
diversity and Euclidean distance give Rao phylogenetic distance.
}
