% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdecorana.R
\name{rdecorana}
\alias{rdecorana}
\alias{scores.rdecorana}
\title{Decorana: R implementation}
\usage{
rdecorana(
  x,
  iweigh = 0,
  iresc = 4,
  ira = 0,
  mk = 26,
  short = 0,
  before = NULL,
  after = NULL
)

\method{scores}{rdecorana}(x, display = "sites", choices = 1:4, origin = FALSE, ...)
}
\arguments{
\item{x}{\code{rdecorana} result object.}

\item{iweigh}{Downweighting of rare species (0: no).}

\item{iresc}{Number of rescaling cycles (0: no rescaling).}

\item{ira}{Type of analysis with numeric values \code{0} classic
detrending by segments, \code{1} orthogonal for standard
correspondence analysis a.k.a. reciprocal averaging, \code{2}
cubic detrending by third degree polynomials, \code{3}
smooth detrending using \code{\link{loess}}, \code{4}
\code{\link{smooth.spline}} with \code{cv = TRUE}.}

\item{mk}{Number of segments in classic detrending.}

\item{short}{Shortest gradient to be rescaled.}

\item{before, after}{Definition of Hill's piecewise transformation.}

\item{display}{Scores for \code{"sites"} or \code{"species"}.}

\item{choices}{Axes to be returned.}

\item{origin}{Return centred scores.}

\item{...}{Other arguments passed to functions.}
}
\value{
Currently returns a list of elements \code{evals} of
    Decorana values, with \code{rproj} and \code{cproj} of scaled
    row and column scores.

Row and column scores and the convergence criterion which
    for orthogonal CA is the eigenvalue.
}
\description{
Similar to Fortran implementation in vegan, but all in R.
}
\details{
This function mimics \CRANpkg{vegan} function
\code{\link[vegan]{decorana}}, but is written in \R{}. It is
slower, and does not have all features and support of the
\pkg{vegan} function, and there usually is no need to use this
function for data analysis. The function serves two
purposes. First, as an \R{} function it is easier to inspect the
algorithm than from C or Fortran code. Second, it is more hackable,
and it is easier to develop new features, change code or replace
functionality than in the compiled code. For instance, it would be
trivial to add Detrended Constrained Correspondence Analysis, but
this would be impossible without extensive changes in Fortran in
the \pkg{vegan} function.

The main function \code{rdecorana} performs janitorial tasks, and
for detrended CA it has basically only two loops. The first loop
finds the next correspondence analysis axis for species and sites,
and the second loop performs optional (if \code{iresc > 0} and
\code{ira != 1}) nonlinear rescaling of that axis.

Two non-exported functions are used for finding the correspondence
analysis axis: \code{transvu} runs one step of reciprocal averaging
and the step is repeated so many times that the criterion value
converges. Function \code{transvu} calls \code{detrend} that
performs detrending. The default is to use Hill's non-linear
rescaling on \code{mk} segments, but this can be substituted with
orthogonalization for classic correspondence analysis,
a.k.a. reciprocal averaging. It is also easy to add new alternative
detrending methods that are not a part of
\code{\link[vegan]{decorana}} (see function \code{natto:::detrend}
for examples).  Detrending against previous axis is done one axis
by time starting from first, and then going again down to the
first. For axis 4 the order of detrendings is 1, 2, 3, 2, 1 (in
orthogonal correspondence analysis with \code{ira=1} the going down
is skipped). The criterion value is the one that \pkg{vegan}
\code{\link[vegan]{decorana}} calls \sQuote{Decorana values}: for
orthogonal CA it is the eigenvalue, but for detrended CA it is a
combination of eigenvalues and strength of detrending.

Non-linear rescaling is performed by function \code{stretch} that
calls two functions: \code{segment} to estimate the dispersion of
species, and \code{smooth121} to smooth those estimates on segments
(that number of segments is selected internally and \code{mk} has
no influence). The criterion values for community data \eqn{x} with
species and site scores \eqn{v, u} are based on
\eqn{x_ij(u_i-v_j)^2}{x[ij] * (u[i]-v[j])^2} summarized over sites
\eqn{i}. Site scores \eqn{u} are weighted averages of species
scores \eqn{v}, and therefore species scores should be dispersed
symmetrically around corresponding site scores and the statistic
describes their weighted dispersion. The purpose of rescaling is to
make this dispersion 1 all over the axis. The procedure is
complicated, and it is best inspected from the code (which is
commented).
}
\examples{
data(spurn)
(mod <- rdecorana(spurn))
## current detrending methods
natto:::detrend

if (require(vegan)) {
## compare to decorana
decorana(spurn)
## Comparison of detrending choices
op <- par(mfrow=c(2,2), mar=c(4,4,1,1))
ordiplot(rdecorana(spurn)) # default: DCA
ordiplot(rdecorana(spurn, ira=2)) # cubic DCA
ordiplot(rdecorana(spurn, ira=3)) # smooth DCA
ordiplot(rdecorana(spurn, ira=4)) # smoothing spline with CV
par(op)
}

}
\seealso{
For serious purposes, use \code{\link[vegan]{decorana}} in
    \pkg{vegan}. Function \code{\link{gradrescale}} rescales
    observed gradients similarly as \code{decorana} rescales axes.
}
\author{
Jari Oksanen
}
