% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdca.R
\name{sdca}
\alias{sdca}
\title{Smoothly Detrended Correspondence Analysis}
\usage{
sdca(Y, pairwise = FALSE, ...)
}
\arguments{
\item{Y}{Input data.}

\item{pairwise}{Detrend axis \eqn{k} separately for each previous
axis in order \eqn{1 \dots k-1 \dots 1}. This only influences
axes 3 and 4.}

\item{\dots}{Other arguments (passed to \code{\link{loess}}).}
}
\description{
Function \code{sdca} is similar to \code{\link[vegan]{decorana}},
but instead of detrending by segments it uses \code{\link{loess}}
for smooth non-linear detrending.
}
