% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slopediff.R
\name{slopediff}
\alias{plot.slopediff}
\alias{slopediff}
\alias{slopediff2}
\alias{summary.slopediff}
\title{Compare Slopes of Mantel Regression Within and Across Blocks}
\usage{
slopediff(ydist, xdist, block, nperm = 999)

slopediff2(ydist, xdist, block, nperm = 999)

\method{plot}{slopediff}(x, col = 1, obsline = TRUE, ...)

\method{summary}{slopediff}(object, ...)
}
\arguments{
\item{ydist,xdist}{Dependent and indpendent dissimilarities.}

\item{block}{Blocking of sites.}

\item{nperm}{Number of simple permutations.}

\item{x}{Result output from \code{slopediff} or \code{slopediff2}.}

\item{col}{Plotting colour.}

\item{obsline}{Draw line on observed statistic in the
\code{\link{density}} plot. .}

\item{object}{Result object from \code{slopediff} or \code{slopediff2}.}
}
\description{
Functions fit linear regression on distances within blocks and
across blocks and compere their slopes using permutation tests.
}
\details{
Functions are intended to analyse the differences of linear Mantel
regression within and across blocks. Typically the dependent
dissimilarities (\code{ydist}) are based on community data,
independent dissimilarities (\code{xdist}) are based on
environmental data, and \code{block} defines subsets of data, such
as geographical subareas.

Functions \code{slopediff} and \code{slopediff2} differ in their
permutation models. Function \code{slopediff} permutes the
classification vector defining \code{block}s, and keeps the
dependent (\code{ydist}) and independent (\code{xdist}) pairs of
dissimilarities fixed. Function \code{slopediff2} permutes
dependent data, and keeps \code{block}s and dependent distances
fixed.
}
\author{
Jari Oksanen
}

