\name{spurn}
\alias{spurn}
\docType{data}
\title{
  Dune Scrub Vegetation in Spurn Point, Yorkshire
}

\description{ Shimwell (1971) used the Spurn Point Dune Scrub data as
  an example well suited for Braun-Blanquet tabular rearrangement
  method.  }

\usage{data("spurn")}
\format{
  A data frame with 20 sample plots on the following 40 species.
  \describe{
    \item{\code{Elaerham}}{\emph{Elaeagnus rhamnoides} (as \emph{Hippophae})}
    \item{\code{Jacovulg}}{\emph{Jacobaea vulgaris} (as \emph{Senecio jacobaea})}
    \item{\code{Soladulc}}{\emph{Solanum dulcamara}}
    \item{\code{Rubufrut}}{\emph{Rubus fruticosus} s.lat.}
    \item{\code{UrtidioiA}}{\emph{Urtica dioica}}
    \item{\code{Rumecris}}{\emph{Rumex crispus}}
    \item{\code{ClayperfB}}{\emph{Claytonia perfoliata} (as \emph{Montia})}
    \item{\code{StelmediB}}{\emph{Stellaria media}}
    \item{\code{FestrubrC}}{\emph{Festuca rubra}}
    \item{\code{ElymrepeC}}{\emph{Elymus repens} (as \emph{Agropyron})}
    \item{\code{AmmoarenC}}{\emph{Ammophila arenaria}}
    \item{\code{Soncarve}}{\emph{Sonchus arvensis}}
    \item{\code{Ononspin}}{\emph{Ononis spinosa} (as \emph{O. repens})}
    \item{\code{Galiveru}}{\emph{Galium verum}}
    \item{\code{Calysold}}{\emph{Calystegia soldanella}}
    \item{\code{PoapratC}}{\emph{Poa pratensis}}
    \item{\code{Agrostol}}{\emph{Agrostis stolonifera}}
    \item{\code{RanubulbC}}{\emph{Ranunculus bulbosus}}
    \item{\code{PlanlancC}}{\emph{Plantago lanceolata}}
    \item{\code{Verocham}}{\emph{Veronica chamaedrys}}
    \item{\code{Epilangu}}{\emph{Epilobium angustifolium} (as \emph{Chamaenerion})}
    \item{\code{CeraglomB}}{\emph{Cerastium glomeratum} (as \emph{C. vulgatum})}
    \item{\code{Sambnigr}}{\emph{Sambucus nigra}}
    \item{\code{CirsvulgB}}{\emph{Cirsium vulgare}}
    \item{\code{Heraspho}}{\emph{Heracleum sphondylium}}
    \item{\code{Inulcony}}{\emph{Inula conyza}}
    \item{\code{CardhirsB}}{\emph{Cardamine hirsuta}}
    \item{\code{Hyporadi}}{\emph{Hypochaeris radicata}}
    \item{\code{Arrhelat}}{\emph{Arrhenatherum elatius}}
    \item{\code{Soncaspe}}{\emph{Sonchus asper}}
    \item{\code{EurhpraeA}}{\emph{Eurhynchium praelongum}}
    \item{\code{Hypncupr}}{\emph{Hypnum cupressiforme}}
    \item{\code{Bracruta}}{\emph{Brachythecium rutabulum}}
    \item{\code{GeasfornB}}{\emph{Geastrum fornicatum}}
    \item{\code{BracalbiC}}{\emph{Brachythecium albicans}}
    \item{\code{Bryuincl}}{\emph{Bryum inclinatum}}
    \item{\code{Syntrura}}{\emph{Syntrichia ruralis} (as \emph{Tortula ruraliformis})}
    \item{\code{Cladranf}}{\emph{Cladonia rangiformis}}
    \item{\code{Bovinigr}}{\emph{Bovista nigrescens}}
    \item{\code{Lophhete}}{\emph{Lophocolea heterophylla}}
  }
  
  The species abundances were visually assessed with Braun-Blanquet
  scale values +, 1 \dots 5 which are presented in numeric scale 1
  \dots 6 in the data set.

  Shimwell (1971) used the data set to demonstrate the Braun-Blanquet
  tabular rearrangement method. He divided the data into three
  classification units marked as A, B and C, and removed one data set
  marked as X. The sample plots are named with consecutive numbers
  after a letter corresponding to these classes. The diagnostic taxa
  have a corresponding upper case letter after their name.

}

\source{ Shimwell, D. W. (1971) \emph{Description and Classification
  of Vegetation (Sidgwick & Jackson), Table 44.  }

}
\examples{
data(spurn)
}
\keyword{datasets}
