% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdiag.R
\name{subdiag}
\alias{subdiag}
\title{Extract Subdiagonal from Dissimilarities or from a Square Matrix}
\usage{
subdiag(x)
}
\arguments{
\item{x}{Square matrix or a \code{\link{dist}} object.}
}
\description{
Extract Subdiagonal from Dissimilarities or from a Square Matrix
}
