% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxondist.R
\name{taxondist}
\alias{taxondist}
\title{Return Gammaplus and Thetaplus Elements of Clarke's Taxonomic
Community Dissimilarity}
\usage{
taxondist(x, d, method = c("gamma", "theta"), dmax)
}
\arguments{
\item{x}{Community data; will be treated as binary presence/absence
matrix.}

\item{d}{Taxonomic, phylogenetic or other dissimilarities among
species (columns of \code{x}).}

\item{method}{Type of returned dissimilarity index.}

\item{dmax}{Scale dissimilarities by \code{dmax} (if \code{dmax >
max(d)}) or truncate dissimilarities at \code{dmax} (if
\code{dmax < max(d)}).}
}
\value{
Clarke's taxonomic dissimilarity index as defined in
    \code{type}.
}
\description{
Return Gammaplus and Thetaplus Elements of Clarke's Taxonomic
Community Dissimilarity
}
