% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.prcomp.rda.R
\name{as.prcomp}
\alias{as.prcomp}
\alias{as.prcomp.rda}
\title{Cast vegan::rda Result to base::prcomp}
\usage{
as.prcomp(x)

\method{as.prcomp}{rda}(x)
}
\arguments{
\item{x}{An unconstrained \code{\link[vegan]{rda}} result object.}
}
\description{
Function casts a result object of unconstrained
\code{\link[vegan]{rda}} to a \code{\link{prcomp}} result object.
}
