% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{bayesjaccard}
\alias{bayesjaccard}
\title{Community Dissimilarity as Expected or Sampled Beta Variate}
\usage{
bayesjaccard(x, method = c("rbeta", "expected"))
}
\arguments{
\item{x}{Community data, will be treated as binary.}

\item{method}{Dissimilarity as a random sample from Beta
distribution or as its expected value.}
}
\description{
Community Dissimilarity as Expected or Sampled Beta Variate
}
