% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{bjdbrda}
\alias{bjdbrda}
\alias{scores.bjdbrda}
\alias{plot.bjdbrda}
\alias{boxplot.bjdbrda}
\title{Multiple dbRDA from Beta Distributed Jaccard Dissimilarity}
\usage{
bjdbrda(formula, data, n = 100, ...)

\method{scores}{bjdbrda}(
  x,
  choices = 1:2,
  display = c("wa", "lc", "bp", "cn"),
  scaling = "species",
  const,
  expected = TRUE,
  ...
)

\method{plot}{bjdbrda}(
  x,
  choices = 1:2,
  wa = "p",
  lc = "n",
  cn = "hull",
  bp = "wedge",
  wa.par = list(),
  lc.par = list(),
  cn.par = list(),
  bp.par = list(),
  scaling = "species",
  type = "t",
  ...
)

\method{boxplot}{bjdbrda}(
  x,
  kind = c("eigen", "correlation"),
  points = "red",
  pch = 16,
  xlab = "Axis",
  ylab,
  ...
)
}
\arguments{
\item{formula, data}{Model definition of type \code{Y ~ Var1 + Var2,
data = X}, where \code{Y} is dependent community data (handled
as binary data), \code{Var1} and \code{Var2} are independent
(explanatory) variables found in data frame \code{X}. See
\code{\link[vegan]{dbrda}} for further information.}

\item{n}{Number of Jaccard dissimilarity matrices sampled from Beta
distribution.}

\item{...}{Other parameters passed to functions; passed
\code{\link[vegan]{dbrda}} in \code{bjdbrda}, ignored in
\code{scores}.}

\item{x}{\code{bjdbrda} result object.}

\item{choices}{Selected ordination axes.}

\item{display, scaling, const}{Kind of scores, the scaling of scores
(axes), and scaling constant with similar definitions as in
\code{\link[vegan]{scores.rda}}.}

\item{expected}{Return scores of the expected ordination instead of
ordination based on random samples from Jaccard dissimilarity.}

\item{wa, lc, cn, bp}{Display of corresponding scores. \code{"n"}
skips the score, \code{"p"} and \code{"t"} use points or text
for the expected score, and other shapes define
\code{\link{bjpolygon}} or \code{\link{bjstars}} shape used for
sampled random scores.}

\item{wa.par, lc.par, cn.par, bp.par}{List of arguments to modify the
plotting parameters of the corresponding shape.}

\item{type}{Add \code{"t"}ext or \code{"p"}oint for the expected
score to a shape of scatter of random shapes.}

\item{kind}{Draw boxplots of eigenvalues or pairwise axis correlations
for each random sample.}

\item{points, pch}{Add points of given color and shape to the
eigenvalue boxplot.}

\item{xlab, ylab}{Change labelling of axes in boxplot.}
}
\description{
Multiple dbRDA from Beta Distributed Jaccard Dissimilarity
}
