% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{bjNMDS}
\alias{bjNMDS}
\alias{plot.bjnmds}
\title{Multiple NMDS Ordinations from Beta Distributed Jaccard Dissimilarity}
\usage{
bjNMDS(
  x,
  n = 100,
  trymax = 500,
  maxit = 1000,
  smin = 1e-04,
  sfgrmin = 1e-07,
  sratmax = 0.999999,
  parallel = 2,
  trace = FALSE,
  ...
)

\method{plot}{bjnmds}(
  x,
  choices = 1:2,
  kind = c("hull", "ellipse", "wedge", "star"),
  keep = 0.9,
  type = "t",
  ...
)
}
\arguments{
\item{x}{Community data to be analysed and treated as binary
(\code{bjNMDS}) or object to be plotted (\code{plot}).}

\item{n}{Number of random samples of Beta Distribution.}

\item{trymax}{Maximum number of random starts in
\code{\link[vegan]{metaMDS}}.}

\item{maxit, smin, sfgrmin, sratmax}{Convergence parameters in
\code{\link[vegan]{monoMDS}}.}

\item{parallel}{Number of parallel tries in \code{\link[vegan]{metaMDS}}.}

\item{trace}{Trace iterations in \code{\link[vegan]{metaMDS}}.}

\item{...}{Other parameters passed to functions.}

\item{choices}{Axes to be plotted.}

\item{kind}{Shape to be plotted to show the scatter of coordinates
of sampled distances; see \code{\link{bjpolygon}} for details.}

\item{keep}{Proportion of points to be enclosed by shape; see
\code{\link{peelhull}} and \code{\link{peelellipse}}.}

\item{type}{Type of the plot: \code{"t"}ext, \code{"p"}oints or
\code{"n"}one.}
}
\value{
A \code{\link[vegan]{metaMDS}} result object amended with item
\code{rscores} that is a three-dimensional array of coordinates
from random \code{\link{bayesjaccard}} ordinations.
}
\description{
Function performs \code{\link[vegan]{metaMDS}} on expected Beta
Jaccard dissimilarity with multiple starts and then a number of
\code{\link[vegan]{monoMDS}} runs on random Beta Jaccard
dissimilarities starting from the expected configuration, and
Procrustes rotates the random solutions to the expected one.
}
\details{
Current function is designed for visual inspection of random
variation of ordination, and no numerical analysis functions are
(yet) available. The \code{plot} can show the scatter of points as
convex hulls or ellipsoid hulls containing a given proportion of
random points, or as stars that connect the expected point to
randomized ones. With option \code{"wedge"} the convex hull is
extended to include the observed point if necessary. See
\code{\link{bjpolygon}} and \code{\link{bjstars}} for technical
details.

Missing functionality includes adding species scores, adding fitted
environmental vectors and factors and numeric summaries of the
randomization results.
}
\examples{

data(spurn)
m <- bjNMDS(spurn)
plot(m, keep = 0.607, col = "skyblue")
}
