% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{bjpolygon}
\alias{bjpolygon}
\alias{bjstars}
\title{Shapes to Display Scatter of Points in Multiple Ordinations}
\usage{
bjpolygon(
  xarr,
  x0,
  keep = 0.9,
  kind = c("hull", "ellipse"),
  criterion = "area",
  linetopoint = TRUE,
  col = "gray",
  alpha = 0.3,
  observed = TRUE,
  type = c("t", "p", "n"),
  ...
)

bjstars(xarr, x0, keep = 0.9, col = "gray", type = c("t", "p", "n"), ...)
}
\arguments{
\item{xarr}{3-D array of coordinates of sampling units times two
axes by random samples.}

\item{x0}{2-D array of sampling units times two axes treated as
constant for all samples in \code{xarr}.}

\item{keep}{Proportion of points enclosed in the shape; passed to
\code{\link{peelhull}} or \code{\link{peelellipse}}.}

\item{kind}{Shape is either a convex hull (\code{\link{peelhull}})
or ellipse \code{\link{peelellipse}} enclosing \code{keep}
proportion of \code{xarr} points.}

\item{criterion}{Criterion to remove extreme points from the hull
in \code{\link{peelhull}} and \code{\link{peelellipse}}.}

\item{linetopoint}{Draw line from the centre of the shape to the
coordinates in \code{x0}.}

\item{col}{Colour of the shape; can be a vector of colours.}

\item{alpha}{Transparency of shapes; 0 is completely transparent,
and 1 is non-transparent.}

\item{observed}{After finding the convex hull, extend hull to
enclose fixed point \code{x0}.}

\item{type}{Mark coordinate of \code{x0} using \code{"t"}ext,
\code{"p"}oint or \code{"n"}one.}

\item{...}{Other parameters passed to to marker of \code{type}.}
}
\value{
Functions return invisibly \code{NULL}.
}
\description{
These are low-level functions that are used by \code{plot}
functions to draw shapes enclosing a specified proportion of
randomized points, or connecting certain proportion of them to the
reference points. The functions are not usually called directly by
users, but the plots can be modified with described arguments.
}
\details{
Functions require output of three-dimensional array \code{xarr} of
randomized scores, where the last dimension is for the random
samples, and a two-dimensional matrix of references scores
\code{x0}.

Function \code{bjpolygon} uses \code{\link{peelhull}} or
\code{\link{peelellipse}} to find a convex hull
(\code{\link{chull}}) or an ellipsoid hull
(\code{\link[cluster]{ellipsoidhull}}) containing \code{keep}
proportion of points. The reference coordinates can also be added
to the plot, either as point or a text label
(\code{\link{ordiellipse}}) and optionally connected to the shape
centre (not to the centre of points). With argument \code{observed}
the convex hull can be extended to include the reference point when
that is outside the hull; this is used to draw wedges for arrows
using the origin as the reference point.

Function \code{bjstars} connects the reference point to \code{keep}
proportion of closest points. If the reference point is not
specified, the centre of points prior to selection will be used.
}
\seealso{
\code{\link{peelhull}}, \code{\link{peelellipse}},
    \code{\link{polygon}} for basic functions and
    \code{\link{plot.bjnmds}} and \code{\link{plot.bjnmds}} and
    \code{\link{plot.bjdbrda}} for programmatic interface.
}
