% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caxNMDS.R
\name{caxNMDS}
\alias{caxNMDS}
\alias{caxNMDSengine}
\alias{plot.caxNMDS}
\alias{stressplot.caxNMDS}
\title{Constrained Non-metric Multidimensional Scaling}
\usage{
caxNMDS(formula, data, k = 2, method = "BFGS")

caxNMDSengine(D, u, k, mm, method = "BFGS")

\method{plot}{caxNMDS}(x, ...)

\method{stressplot}{caxNMDS}(object, dis, p.col = "blue", l.col = "red", lwd = 2, ...)
}
\arguments{
\item{formula}{Model formula where the left-hand-side is a distance
structure for depenedent (community) dissimilarity and
right-hand-side specifies the contraints.}

\item{data}{Data frame of constraints.}

\item{k}{Number of dimensions in NMDS.}

\item{method}{Optimization method used in \code{\link{optim}}.}

\item{D}{Dissimilarities.}

\item{u}{Initial configuration used for starting values in
\code{\link{optim}}. Solution from metric scaling
(\code{\link[vegan]{wcmdscale}}) is used if this is missing.}

\item{mm}{\code{\link{model.matrix}} of constraints.}

\item{x}{Result object from \code{caxNMDS}.}

\item{\dots}{Other arguments passed to graphical functions.}

\item{object}{Result object from \code{caxNMDS}.}

\item{dis}{Original community dissimilarities used in
\code{caxNMDS} (must be supplied).}

\item{p.col}{Colour of points}

\item{l.col, lwd}{Colour and width of expected values form isotonic
regression.}
}
\value{
\code{caxNMDSengine} returns the \code{\link{optim}} result
    object and adds items \code{points} for ordination scores,
    \code{stress} for goodness of fit (\sQuote{stress}), and
    \code{coefficients} for a matrix of regression coefficients of
    constraints. Function \code{caxNMDS} returns these added items
    plus \code{formula}, \code{call} and fitted vectors and factor
    centroids of constraints in element \code{ef} with \pkg{vegan}
    function \code{\link[vegan]{envfit}}.
}
\description{
Constrained non-metric Multidimensional scaling is a monotonous
(isometric) regression (\code{\link[stats]{isoreg}}) of observed
(community) dissimilarities on dimensions that are linear
combinations of constraints.
}
\details{
Ordinary non-constrained NMDS is a monotonous (isometric)
regression (\code{\link[stats]{isoreg}}) of observed (community)
dissimilarities on ordination scores. Constrained NMDS is similar,
but restricts the ordination scores to linear combinations of
constraining variables.

Function uses standard \code{\link[stats]{optim}} to find the
regression coefficients of constraints. The regression coefficients
are the primary result of optimization, and ordination scores are
found applying regression coefficients on the model matrix of
constraints. The criterion variable to be minimized is
\code{stress} which is identical to stress in non-constrained NMDS
(\code{\link[vegan]{monoMDS}}, \code{\link[vegan]{metaMDS}} in
\pkg{vegan}). Function uses \dQuote{weak} ties, or allow breaking
tied values in monotonous regression.

Function \code{caxNMDSengine} is the numerical core of the
method. It needs the input dissimilarities and model matrix
(without constant intercept). Function \code{caxNMDS} is more
user-friendly interface with model formula for constraints, and
also adds fitted vectors of continuous constraints and level
centroids of factor constraints. These are similar to biplot scores
and constraints in constrained ordination methods such as
\pkg{vegan} \code{\link[vegan]{rda}} and
\code{\link[vegan]{dbrda}}.

Function \code{\link{cdisNMDS}} in this package provides an
alternative method based on constrained dissimilarities
(\code{\link{distconstrain}}) instead of constrained axes. The
current \code{caxNMDS} function seems to be more robust.

This is a proof-of-concept function based on the idea of Bert van
der Veen.
}
\note{
Function \code{caxNMDSengine} can be used as \code{engine} in
    \code{\link[vegan]{metaMDS}} in \pkg{vegan} development version
    (2.8-0; not yet released).
}
\examples{
data(mite, mite.env, package = "vegan")
dis <- vegan::vegdist(mite)
mod <- caxNMDS(dis ~ WatrCont + SubsDens + Shrub + Topo,
  data = mite.env)
plot(mod$ef)
coef(mod)
mod$ef
plot(mod)
stressplot(mod, dis, cex = 0.3)
}
\seealso{
\code{\link{cdisNMDS}} which provides an alternative and
    worse alternative that is based on monotonous (isometric)
    regression of constrained dissimilarities
    \code{\link{distconstrain}}.
}
\author{
Jari Oksanen and Bert van der Veen.
}
