% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caxNMDS.R
\name{caxNMDS}
\alias{caxNMDS}
\alias{caxNMDSengine}
\title{Constrained Non-metric Multidimensional Scaling}
\usage{
caxNMDS(formula, data, k = 2, u, method = "BFGS")

caxNMDSengine(D, u, k, mm, method = "BFGS")
}
\arguments{
\item{formula}{Model formula where the left-hand-side is a distance
structure for depenedent (community) dissimilarity and
right-hand-side specifies the contraints.}

\item{data}{Data frame of constraints.}

\item{k}{Number of dimensions in NMDS.}

\item{u}{Initial configuration used for starting values in
\code{\link{optim}}.}

\item{method}{Optimization method used in \code{\link{optim}}.}

\item{D}{Dissimilarities.}

\item{mm}{\code{\link{model.matrix}} of constraints.}
}
\description{
Constrained non-metric Multidimensional scaling is a monotonous
(isometric) regression (\code{\link[stats]{isoreg}}) of observed
(community) dissimilarities on dimensions that are linear
combinations of constraints.
}
\details{
This is a proof-of-concept function based on the idea of Bert van
der Veen.
}
\note{
Function \code{caxNMDS} can be used as \code{engine} in
    \code{\link[vegan]{metaMDS}} in \pkg{vegan} development version
    (2.8-0; not yet released).
}
\examples{
data(mite, mite.env, package = "vegan")
dis <- vegan::vegdist(mite)
mod <- caxNMDS(dis ~ WatrCont + SubsDens + Shrub + Topo,
  data = mite.env)
vegan::ordiplot(mod, display = "sites")
plot(mod$ef)
coef(mod)
mod$ef

}
\seealso{
\code{\link{cdisNMDS}} which provides an alternative and
    worse alternative that is based on monotonous (isometric)
    regression of constrained dissimilarities
    \code{\link{distconstrain}}.
}
\author{
Jari Oksanen and Bert van der Veen.
}
