% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caxNMDS.R
\name{caxNMDS}
\alias{caxNMDS}
\title{Constrained Non-metric Multidimensional Scaling}
\usage{
caxNMDS(formula, data, k = 2)
}
\description{
Constrained non-metric Multidimensional scaling is a monotonous
(isometric) regression (code{\link[stats]{isoreg}}) of observed
(community) dissimilarities on dimensions that are linear
combinations of constraints.
}
\details{
This is a proof-of-concept function based on the idea of Bert van
der Veen.
}
\examples{
data(mite, mite.env, package = "vegan")
mod <- caxNMDS(vegdist(mite) ~ WatrCont + SubsDens + Shrub + Topo,
  data = mite.env)
ordiplot(mod, display = "sites")
plot(mod$ef)
coef(mod)
mod$ef
}
\seealso{

}
