% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caxNMDS.R
\name{caxNMDS}
\alias{caxNMDS}
\title{Constrained Non-metric Multidimensional Scaling}
\usage{
caxNMDS(formula, data, k = 2)
}
\arguments{
\item{formula}{Model formula where the left-hand-side is a distance
structure for depenedent (community) dissimilarity and
right-hand-side specifies the contraints.}

\item{data}{Data frame of constraints.}

\item{k}{Number of dimensions in NMDS.}
}
\description{
Constrained non-metric Multidimensional scaling is a monotonous
(isometric) regression (\code{\link[stats]{isoreg}}) of observed
(community) dissimilarities on dimensions that are linear
combinations of constraints.
}
\details{
This is a proof-of-concept function based on the idea of Bert van
der Veen.
}
\examples{
data(mite, mite.env, package = "vegan")
mod <- caxNMDS(vegdist(mite) ~ WatrCont + SubsDens + Shrub + Topo,
  data = mite.env)
ordiplot(mod, display = "sites")
plot(mod$ef)
coef(mod)
mod$ef
}
\seealso{
\code{\link{cdisNMDS}} which provides an alternative and
    worse alternative that is based on monotonous (isometric)
    regression of constrained dissimilarities
    \code{\link{distconstrain}}.
}
