% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisNMDS.R
\name{cdisNMDS}
\alias{cdisNMDS}
\alias{scores.cdisNMDS}
\alias{plot.cdisNMDS}
\title{Constrained Distance Nonmetric Multidimensional Scaling}
\usage{
cdisNMDS(formula, data, k = 2, add = FALSE)

\method{scores}{cdisNMDS}(x, display = "sites", ...)

\method{plot}{cdisNMDS}(x, display = "sites", type = "p", ...)
}
\arguments{
\item{formula}{Model formula where the left-hand side must be
dissimilarities, and right-hand side the constraints.}

\item{data}{Data frame to find the terms on the right-hand side of
the formula.}

\item{k}{Number of dimensions in NMDS.}

\item{add}{Type of additive constant used to avoid negative squared
distances in \code{\link{distconstrain}}. Either
\code{"lingoes"} or \code{"cailliez"} or \code{FALSE}.}

\item{x}{\code{cdisNMDS} result object.}

\item{display}{Kind of scores to display. In \code{scores} this can
be one or several of \code{"sites"}, \code{"constraints"},
\code{"biplot"}, \code{"centroids"}, or alternative
\code{"all"} for all these. \code{plot} accepts only one
alternative (and no \code{"all"}).}

\item{\dots}{Other arguments passed to graphics functions.}

\item{type}{Either \code{"t"}ext, \code{"p"}oints or \code{"n"}one.}
}
\value{
Function returns an object of class \code{"cdisNMDS"} that
    inherits from \pkg{vegan} function
    \code{\link[vegan]{monoMDS}}. but it has rudimentary
    \code{scores} and \code{plot} methods. The \code{plot} only
    shows one set of scores, and it is best to build graphics with
    pipes (see Examples). The result object also contains biplot
    arrows for continuous constraints and centroids for factor
    constraints. These are based on community ordination (sites)
    instead of constraints, and they are added with
    \code{\link[vegan]{envfit}}. Species scores are not available.
}
\description{
Constraints are estimated as non-metric MDS of constrained
distances (\code{\link{distconstrain}}). These are similar as the
constraints (LC) scores of distance-based RDA
(\code{\link[vegan]{dbrda}}), but with non-metric mapping. Site
ordination is rank-order similar to the constraints and community
points are added to the constraints using \pkg{vegan} function
\code{\link[vegan]{MDSaddpoints}} with original unconstrained
community dissimilarities.
}
\details{
The steps of the algorithm are:
\enumerate{
  \item Find constrained dissimilarities for given constraints
    with \code{\link{distconstrain}}.
  \item Find NMDS of constrained dissimilarities with
    \code{\link[vegan]{metaMDS}}.
  \item Add points to constrained ordination using unconstrained
    community dissimilarities with \code{\link[vegan]{MDSaddpoints}}.
} % end enumerate

The algorithms creates analogous result to constrained eigenvector
ordination distance-based RDA (db-RDA, \code{\link[vegan]{dbrda}}).
The constrained component is similar to linear combination (LC)
scores and the site NMDS analogous to to weighted average (WA) or
site scores of db-RDA. The \code{cdisNMDS} site scores are found by
adding the sites with their community dissimilarities as new points
to the ordination of constraints.  Instead of raw dissimilarities,
both the constrained ordination and the site ordination
("new points") are based on ranks of dissimilarities (with
ties). The ranks give the same result as raw dissimilarities within
one set of dissimilarities (constrained, community), but with them
constrained and community components are rank-order similar.

Although the idea of the algorithm is simple and obvious, the
results are often far from satisfactory. Quite often the community
dissimilarities and constrained dissimilarities differ from each
other so strongly that the site points (community) are expelled
from the constrained configuration and are located at the outskirts
of the ordination instead of being mixed with the constrained
points. The community (site) points and constraints mix well only
when the constraints describe the community dissimilarities
sufficiently well. Function \code{\link{caxNMDS}} in this package
provides an alternative and more robust alternative which
constraints the axes (dimensions) instead of dissimilarities.
}
\section{Warning}{


The function is provided as an exhibit of the algorithm. The
results are often poor, and the function should not be used for any
other purposes than inspecting the method and as an inspiration for
alternative implementations. See \code{\link{caxNMDS}} for more
promising alternative algorithm.
}

\examples{
if (require(vegan)) { # vegan needed for data
data(mite, mite.env)
dis <- canneddist(mite, "chord")
mod <- cdisNMDS(dis ~ WatrCont + SubsDens + Topo + Shrub, mite.env)
print(mod)
plot(mod, type = "p") |>
   points("constraints", pch = 16, col = 2) |>
   text("biplot", col = 4) |>
   text("centroids", col = 4)
legend("topleft", c("LC Scores", "NMDS Scores", "Constraints"),
   col = c(2,1,NA), pch = c(16,1,NA), text.col = c(1,1,4))
}

}
\seealso{
\code{\link{distconstrain}}, \code{\link[vegan]{dbrda}},
    \code{\link[vegan]{monoMDS}},
    \code{\link[vegan]{MDSaddpoints}}.  Function
    \code{\link{caxNMDS}} is a more robust alternative that
    constraints the dimensions (\dQuote{LC scores}) instead of
    dissimilarities.
}
\author{
Jari Oksanen
}
