% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesjaccard.R
\name{clsupport}
\alias{clsupport}
\alias{clsets}
\title{Support of Branches of Hierarchical Clustering from Beta Jaccard}
\usage{
clsupport(x, n = 1000, method = "average", softmatch = FALSE, plot = TRUE, ...)

clsets(hclus)
}
\arguments{
\item{x}{Community data, will be treated as binary.}

\item{n}{Number of random samples from Beta distribution.}

\item{method}{Clustering method, passed to \code{\link{hclust}}.}

\item{softmatch}{Estimate cluster similarity as Jaccard similarity
or as a hard exact match between two clusters.}

\item{plot}{Plot the \code{\link{hclust}} dendrogram from expected
Jaccard dissimilarity with each brand labelled with support.}

\item{...}{Other parameters; passed to \code{\link{plot.hclust}}.}

\item{hclus}{\code{\link{hclust}} result object.}
}
\description{
Support of Branches of Hierarchical Clustering from Beta Jaccard
}
