% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cNMDS.R
\name{cNMDS}
\alias{cNMDS}
\alias{scores.cNMDS}
\alias{plot.cNMDS}
\title{Constrained Nonmetric Multidimensional Scaling}
\usage{
cNMDS(formula, data, k = 2, add = FALSE)

\method{scores}{cNMDS}(x, display = "sites", ...)

\method{plot}{cNMDS}(x, display = "sites", type = "p", ...)
}
\arguments{
\item{formula}{Model formula where the left-hand side must be
dissimilarities, and right-hand side the constraints.}

\item{data}{Data frame to find the terms on the right-hand side of
the formula.}

\item{k}{Number of dimensions in NMDS.}

\item{add}{Type of additive constant used to avoid negative squared
distances in \code{\link{distconstrain}}. Either
\code{"lingoes"} or \code{"cailliez"} or \code{FALSE}.}

\item{display}{Kind of scores to display. Can be one or several of
\code{"sites"}, \code{"constraints"}, \code{"biplot"},
\code{"centroids"}, or alternative \code{"all"} for all these.}

\item{type}{Either \code{"t"}ext, \code{"p"}oints or \code{"n"}one.}
}
\description{
Adds points based on community dissimilarities on the NMDS
based on constrained dissimilarities.
}
\details{
The steps of algorithm are:
\enumerate{
  \item Find constrained dissimilarities for given constraints
    with \code{\link{distconstrain}}.
  \item Find NMDS of constrained dissimilarities with
    \code{\link[vegan]{monoMDS}} using metric scaling as initial
    configuration.
  \item Add points to constrained ordination using unconstrained
    community dissimilarities with \code{\link[vegan]{MDSaddpoints}}.
} % end enumerate
}
