% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginkgogram.R
\name{ginkgogram}
\alias{ginkgogram}
\title{Plot Dendrogram with Leaves as Fans}
\usage{
ginkgogram(x, labels = NULL, check = TRUE, axes = TRUE,
  frame.plot = FALSE, ann = TRUE, main = "Cluster Ginkgogram",
  sub = NULL, xlab = NULL, ylab = "Height", w, col, leaf = Inf,
  ...)
}
\arguments{
\item{x}{An object of the type produced by \code{\link{hclust}}.}

\item{labels}{A character vector of labels for te leaves of the
tree.}

\item{check}{Logical indicating if the \code{x} should be checked
for validity.}

\item{axes, frame.plot, ann}{Logical flags as in
\code{\link{plot.default}}.}

\item{main, sub, xlab, ylab}{Character strings to replace default
annotation.}

\item{w}{Weights of leaves.}

\item{col}{Colour of the fill of leaves.}

\item{leaf}{Maximum height of leaf triangle. Leaf height is
smaller of this value and the height of the branch, and the
default value (\code{Inf}) always draws leaves to the
root. Special value \code{TRUE} will take the smallest branch
height as the maximum height of each leaf.}

\item{\dots}{Further graphical arguments}
}
\description{
Function is similar to standard \code{\link{hclust}}, but the
leaves are drawn as fans with base proportional to weights (or
sizes) of leaves.
}
\details{
Function \code{ginkgogram} has two new arguments to
    \code{\link{hclust}}: \code{w} for weights that give the base
    width of leaves, and \code{col} that gives the colour of the
    fill of leaves. The weights \code{w} must be numeric and
    non-negative, and zero weight guarantees that the leaf is drawn
    as a simple line similarly as in \code{\link{hclust}}.
}
\examples{
## need vegan data sets
if (require(vegan)) {
data(dune, dune.phylodis, dune.taxon)
cl <- hclust(dune.phylodis)
ginkgogram(cl, w = colMeans(dune), col = factor(dune.taxon$Superorder))
## limit heights of the polygons to the Cretaceous
ginkgogram(cl, w = colMeans(dune), col = factor(dune.taxon$Superorder),
    leaf = 65)
}
}
