% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradrescale.R
\name{gradrescale}
\alias{gradrescale}
\title{Rescale Ecological Gradient to Constant Community Heterogeneity}
\usage{
gradrescale(grad, comm)
}
\arguments{
\item{grad}{Observed numeric gradient.}

\item{comm}{Community data matrix.}
}
\value{
Rescaled gradient scaled in units of heterogeneity.
}
\description{
Rescale Ecological Gradient to Constant Community Heterogeneity
}
\details{
Observed environmental gradient is rescaled in similar way as
species axis in \code{\link{rdecorana}} and
\code{\link[vegan]{decorana}}. The goal is that weighted averages
of species have equal dispersion along the rescaled gradient, but
retain the order of sites along environmental gradient, unlike in
\code{[r]decorana}.

Axis rescaling in \code{[r]decorana} is often poorly
understood. Although rescaling is based on site (sample) scores, it
is primarily performed on species scores, and site (sample) scores
are finally found as weighted averages of rescaled species
scores. With this the rank order of site (sample) scores can
change. \code{[r]decorana} usually performs several sweeps (default
is four), and when new species scores are based on rescaled site
(sample) scores, their rank order can also change from the
original. Changing the order is not acceptable in rescaling observed
gradients, and therefore input gradient is handled like species
scores in \code{[r]decorana} and only one rescaling sweep is made.
}
\examples{
if(require(vegan, quietly=TRUE)) {
data(mite, mite.env)
WatrCont.rescaled <- with(mite.env, gradrescale(WatrCont, mite))
plot(WatrCont.rescaled ~ WatrCont, data=mite.env)
}

}
\seealso{
\code{\link{rdecorana}}, \code{\link[vegan]{decorana}}.
}
\author{
Jari Oksanen
}
