% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mrankdist}
\alias{mrankdist}
\title{Mean Rank Shift Between Pairs of Sites}
\usage{
mrankdist(x)
}
\arguments{
\item{x}{The input data.}
}
\description{
Function calculates the mean rank shift between decreasing rank
orders of species (columns). The averaging is done by the number of
species occurring in both compared sites, and the results are
returned as dissimilarities.
}
\references{
Collins, S.L, Suding, K.N., Cleland, E.E., Batty, M.,
Pennings, S.C., Gross, K.L., Grace, J.B., Gough, L., Fargione,
J.E. & Clark, C.M. (2008). Rank clocks and plant community
dynamics. \emph{Ecology} 89, 3534--3541.
}

