% Generated by roxygen2 (4.0.2): do not edit by hand
\name{occpmap}
\alias{occpmap}
\title{Plot a projected map of species occurrences}
\usage{
occpmap(query, CRS = "+proj=longlat", extent, limit = 1000, pruning,
  clip = TRUE, resolution = "low", border = 1, land = "grey90",
  pch = 16, pcol = 2, annotate = TRUE, ...)
}
\arguments{
\item{query}{Species to be plotted.}

\item{CRS}{Projection used in \code{\link{projectedmap}}.}

\item{extent}{A four element numeric vector of east min, east max,
north min, north max, or other structure accepted by
\code{\link{projectedmap}}.}

\item{limit}{Maximum number of occurrences returned as defined in
\code{\link[spocc]{occ}}.}

\item{pruning}{Remove duplicated points rounded to precision in
latitude and longitude (e.g., \code{pruning = 0.1} keeps only cases
that are not duplicated with 0.1 degree attitude).}

\item{clip}{(logical) Clip map to the occurrences of species
instead of using the original \code{extent}.}

\item{resolution}{Resolution of the map as defined in
\code{\link[rworldmap]{getMap}}.}

\item{border}{Border colour in the map.}

\item{land}{Land colour in the map.}

\item{pch}{Plotting character for points.}

\item{pcol}{Point colour.}

\item{annotate}{Add species name as the title and projection as the
right margin text in the map.}

\item{\dots}{Arguments passed to \code{\link[spocc]{occ}} and
\code{\link{projectedmap}}.}
}
\description{
Function queries species occurrence data with
\code{\link[spocc]{occ}} function and draws a projected map with
these occurrences using \code{\link{projectedmap}}.
}

