% Generated by roxygen2 (4.0.1): do not edit by hand
\name{occpmap}
\alias{occpmap}
\title{Plot a projected map of species occurrences}
\usage{
occpmap(query, CRS = "+proj=longlat", geometry = NULL, limit = 1000,
  pruning, ...)
}
\arguments{
\item{query}{Species to be plotted.}

\item{CRS}{Projection used in \code{\link{projectedmap}}.}

\item{geometry}{As defined in \code{\link[spocc]{occ}}, typically
corner points of the longitude-latitude polygon.}

\item{limit}{Maximum number of occurrences returned as defined in
\code{\link[spocc]{occ}}.}

\item{pruning}{Remove duplicated points rounded to precision in
latitude and longitude (e.g., \code{pruning = 0.1} keeps only cases
that are not duplicated with 0.1 degree attitude).}

\item{\dots}{Arguments passed to \code{\link[spocc]{occ}} and
\code{\link{projectedmap}}}
}
\description{
Function queries species occurrence data with
\code{\link[spocc]{occ}} function and draws a projected map with
these occurrences using \code{\link{projectedmap}}.
}

