% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peelhull.R
\encoding{UTF-8}
\name{peelhull}
\alias{peelhull}
\title{Convex Hull Enclosing a Given Proportion of Points}
\usage{
peelhull(pts, keep = 0.9, criterion = c("area", "distance"))
}
\arguments{
\item{pts}{Coordinates of points, a two-column matrix}

\item{keep}{Proportion of points kept}

\item{criterion}{Criterion to remove a point on the
hull. \code{"area"} removes the point that reduces the area of
the new hull as much as possible, and \code{"distance"} removes
the point that has the largest total distance to all points on
and within the hull.}
}
\value{
A two-column matrix of coordinates of peeled hull defining
    a closed convex hull that can be plotted with
    \code{\link[graphics]{polygon}}.
}
\description{
Function finds a small convex hull enclosing a given proportion of
points. The function works by removing points from the hull one by
one. The points are selected either so that the area of the
remaining hull is reduced as much as possible at every step (de
Smith, Goodrich & Longley 2007), or removing the point that has the
maximal total distance to all remaining points. The function only
works in 2D.
}
\details{
This is preliminary work (but without guarantee of
    progress). However, the current function is such that it can be
    easily plugged into \code{\link[vegan]{ordihull}}
    (\CRANpkg{vegan} package).

The algorithms are naïve, and stepwise removal of single points
does not guarantee smallest possible final hull. Although the area
reduction algorithm was found in literature (de Smith et al. 2007),
the distance criterion often gives smaller final convex hulls.
}
\references{
de Smith, M.J., Goodchild, M.F. & Longley,
    P.A. (2007). \emph{Geospatial analysis: A comprehensive guide
    to principles, techniques and software tools}. Matador.
}
\author{
Jari Oksanen
}
