% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peelhull.R
\encoding{UTF-8}
\name{peelhull}
\alias{peelhull}
\title{Convex Hull Enclosing a Given Proportion of Points}
\usage{
peelhull(pts, keep = 0.9)
}
\arguments{
\item{pts}{Coordinates of points, a two-column matrix}

\item{keep}{Proportion of points kept}
}
\value{
A two-column matrix of coordinates of peeled hull defining
    a closed convex hull (last and first lines are the same).
}
\description{
Function finds a small convex hull enclosing a given proportion of
points. The function works by removing the point from the hull that
gives the largest reduction in the area of remaining hull until a
desired number of points are removed (de Smith, Goodrich & Longley
2007). The function only works in 2D.
}
\details{
This is preliminary work (but without guarantee of
    progress). However, the current function is such that it can be
    easily plugged into \code{\link[vegan]{ordihull}}
    (\CRANpkg{vegan} package).

The algorithm is naïve, and although it was found in literature (de
Smith et al. 2007), it does not guarantee smallest possible hull:
each single step gives largest possible reduction of hull given
previous steps, but there is no guarantee that the sum of single
large reductions gives smallest possible convex hull.
}
\references{
de Smith, M.J., Goodchild, M.F. & Longley,
    P.A. (2007). \emph{Geospatial analysis: A comprehensive guide
    to principles, techniques and software tools}. Matador.
}
\author{
Jari Oksanen
}
