% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarord.R
\name{polarord}
\alias{polarord}
\title{Polar (Bray-Curtis) Ordination}
\usage{
polarord(d, k = 2)
}
\arguments{
\item{d}{Dissimilarities or distances: a \code{\link{dist}} object.}

\item{k}{Number of dimensions.}
}
\value{
Currently the function returns only a matrix of ordination scores,
but this will change.
}
\description{
Polar or Bray-Curtis ordination is a historic ordination method
that could be performed without computers with simple hand
calculations (Bray & Curtis 1957). Ordination axis is found by
selecting two extreme points and projecting all points between
these end points. The current function follows Beals (1984) in
selecting the endpoints, projection of points on axis, and defining
the residual distances for later axes.
}
\details{
The implementation follows McCune & Grace (2002, Chapter 17). The
endpoints are found using the variance-regression method of Beals
(1984). The first endpoint has the highest variance of distances to
other points. This guarantees that the point is at the margin of
the multivariate cloud but is not an outlier, since outliers have
long distances to all points and hence low variance of
distances. The second endpoint has the lowest (most negative)
regression coefficient between distances from the first and second
point to all other points. This selects a point at the margin of
the main cloud of points, opposite to first endpoint. All points
are projected on the axis between the endpoints using Euclidean
geometry, and this gives the scores on a polar ordination
axis. Then the effect of the  axis are removed by calculating
residual distances using Euclidean geometry. Ecological indices are
usually semimetric, and negative residual distances can emerge, but
these are taken as zero in the current function.

Polar ordination is a historical method that is little used today,
but several authors claim that it is a powerful method (see McCune
& Grace 2002). Although the basic operations can be easily
performed by hand or graphically, the later developments of
endpoint selection require more extensive calculations. With modern
numerical utilities, the polar ordination is not faster than metric
multidimensional scaling (\code{\link{cmdscale}},
\code{\link[vegan]{wcmdscale}}).

It is possible to use predefined endpoints instead of automatic
selection. This can be useful for confirmatory analysis (McCune &
Grace 2002). However, this is not (yet) implemented in this
function (but contributions are welcome).
}
\examples{

data(spurn)
## We still need vegan, but work for self-standing case
if(require(vegan)) {
dis <- vegdist(spurn)
ord <- polarord(dis)
ordiplot(ord, type = "t")
}

}
\references{
Beals, E. W. (1984) Bray-Curtis ordination: an effective strategy
for analysis of ecological multivariate data. \emph{Advances in
Ecological Research} 14, 1--55.

Bray, J. R. & Curtis, J. T. (1957) An ordination of the upland
forest communities in southern Wisconsin. \emph{Ecological
Monographs} 27, 325--349.

McCune, B. & Grace, J. B. (2002) \emph{Analysis of Ecological
Communities.} MjM Software Design.
}
