% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raostand.R
\name{raostand}
\alias{raostand}
\title{Standardize Data to Yield Rao Satistics}
\usage{
raostand(x, d)
}
\arguments{
\item{x}{Community data.}

\item{d}{Phylogenetic distances.}
}
\description{
Function standardizes data so that it gives Rao phylogenetic
diversity and Euclidean distance give Rao phylogenetic distance.
}
