% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdecorana.R
\name{rdecorana}
\alias{rdecorana}
\title{Decorana: R implementation}
\usage{
rdecorana(
  x,
  iweigh = 0,
  iresc = 0,
  ira = 0,
  mk = 26,
  short = 0,
  before = NULL,
  after = NULL
)
}
\arguments{
\item{x}{input data matrix.}

\item{iweigh}{Downweighting of rare species (0: no).}

\item{iresc}{Number of rescaling cycles (0: no rescaling).}

\item{ira}{Type of analyis (0: detrended, 1: orthogonal).}

\item{mk}{Number of segments in detrending.}

\item{short}{Shortest gradient to be rescaled.}

\item{before, after}{Definition of Hill's piecewise transformation.}
}
\value{
Currently returns a list of elements \code{evals} of
    Decorana values, with \code{rproj} and \code{cproj} of scaled
    row and column scores.

Row and column scores and the convergence criterion which
    for orthogonal CA is the eigenvalue.
}
\description{
Similar to Fortran implementation in vegan, but all in R.
}
\details{
This function duplicates \CRANpkg{vegan} function
\code{\link[vegan]{decorana}}, and there is no need to use this
function for data analysis. The function serves two
purposes. Firstly, it is written in \R{} to allow easier inspection
of function than compiled code in C and Fortran in
\pkg{vegan}. Secondly, it is more hackable, and easier to develop
new features, change code or replace functionality than in the
compiled code. For instance, it would be trivial to add Detrended
Constrained Correspondence Analysis, but this would be impossible
without extensive changes in Fortran in the \pkg{vegan} function.

Function is experimental, and at the moment it only implements
basic orthogonal and detrended correspondence analysis. It does not
implement many options of \code{\link[vegan]{decorana}}, most
importantly, rescaling is not yet implemented.
}
\examples{
data(spurn)
str(mod <- rdecorana(spurn))
if (require(vegan)) {
## compare to decorana
str(decorana(spurn, iresc = 0))
## ordiplot works
ordiplot(mod, display="sites")
}

}
