% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resassoc.R
\name{resassoc.cca}
\alias{resassoc.cca}
\title{Residual Species Associations in Constrained Ordination}
\usage{
resassoc.cca(x, rank)
}
\arguments{
\item{x}{Constrained ordination result from
\code{\link[vegan]{cca}} or \code{\link[vegan]{rda}}.}

\item{rank}{Number of unconstrained ordination axes that are used
to compute the species associations. If missing, the rank is
chosen by brokenstick distribution
(\code{\link[vegan]{bstick}}).}
}
\value{
Correlation-like association matrix between species with
    added argument \code{"rank"} of the unconstrained data used to
    compute associations.
}
\description{
Function finds residual species association after fitting the
constraints (and conditions) in constrained ordination
(\code{\link[vegan]{cca}}, \code{\link[vegan]{rda}}).
}
\examples{
library(vegan)
data(dune, dune.env)
mod <- cca(dune ~ A1 + Moisture, dune.env)
resassoc.cca(mod)

}
