% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxondist.R
\name{taxondist}
\alias{taxondist}
\title{Return Gammaplus and Thetaplus Elements of Clarke's Taxonomic
Community Dissimilarity}
\usage{
taxondist(x, taxdist)
}
\arguments{
\item{x}{Community data; will be treated as binary presence/absence
matrix.}

\item{taxdist}{Taxonomic, phylogenetic or other dissimilarity
matrix among species (columns of x).}
}
\value{
list of Clarke's Gammaplus and Thetaplus elements as
    distance structures.
}
\description{
Return Gammaplus and Thetaplus Elements of Clarke's Taxonomic
Community Dissimilarity
}
