% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxondist.R
\name{taxondist}
\alias{taxondist}
\title{Return Gammaplus and Thetaplus Elements of Clarke's Taxonomic
Community Dissimilarity}
\usage{
taxondist(x, taxdist, method = c("gamma", "theta"))
}
\arguments{
\item{x}{Community data; will be treated as binary presence/absence
matrix.}

\item{taxdist}{Taxonomic, phylogenetic or other dissimilarity
matrix among species (columns of x).}

\item{method}{Type of returned dissimilarity measure.}
}
\value{
Clarke's taxonomic dissimilarity index as defined in
    \code{type}.
}
\description{
Return Gammaplus and Thetaplus Elements of Clarke's Taxonomic
Community Dissimilarity
}
